% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fs.R
\name{agreed}
\alias{agreed}
\title{Calculate shared features}
\usage{
agreed(df, x)
}
\arguments{
\item{df}{[data.frame] a \code{data.frame} where rows are features and
columns are feature selection methods. Only \code{logical} values for the
cell are allowed.}

\item{x}{[numeric] the selected column positions of \code{df}.}
}
\value{
the number of rows where columns in \code{x} are TRUE, i.e. the
  shared features (rows) between indicated columns (\code{x}, fs_methods).
}
\description{
This function calculates shared features between the indicated columns.
It was meant to be called from \code{plot_venn_fs}.
}
