% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{anova_fs}
\alias{anova_fs}
\title{ANOVA feature selection}
\usage{
anova_fs(input_data, sample_labels, n, cores = 1)
}
\arguments{
\item{input_data}{[data.frame] a \code{data.frame} with samples as rows and
features as columns. Each column is treated as a response (dependent
variable) for an ANOVA model.}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels. It will be used as a predictor (independent
variable) for each ANOVA model. NOTE: this variable should not be included
as a column of \code{input_data}.}

\item{n}{[numeric] the number of features to select.}

\item{cores}{[numeric] number of cores to use for parallel ANOVA fitting.
Default: 1.}
}
\value{
[character] A character vector with the features \code{colnames}
  selected.
}
\description{
Selects features based on ANOVA F-statistic. For each feature, a one-way
ANOVA model is fitted. Resulting p-values are ascending sorted and the
first \code{n} features are selected, based on this sorting.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
features_selected <- anova_fs(input_data, sample_labels, 2)
print(features_selected)

}
