% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\name{apply_limma_filter}
\alias{apply_limma_filter}
\title{Select CpGs by differential methylation}
\usage{
apply_limma_filter(betas, class_labels, limma_select_best)
}
\arguments{
\item{betas}{[matrix] a matrix with normalized beta-values, where rows are
probes and columns are samples.}

\item{class_labels}{[factor] a factor with sample class labels.}

\item{limma_select_best}{[numeric] the number of top differentially
methylated CpGs to select.}
}
\value{
[character] A character vector with the \code{limma_select_best}
  top differentially methylated (sorted by p-value) probe names. If there
  are more than 2 groups, the same number of top CpGs are returned, but
  divided between each one vs the others contrast.
}
\description{
This funtion uses limma to select top differentially methylated probes
between groups (of class labels factor). The number of classes can be 2 or
more, but no paired designs are yet implemented.
}
