% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_methods.R
\name{available_cls_methods}
\alias{available_cls_methods}
\title{Available Classification Methods}
\usage{
available_cls_methods()
}
\description{
This function generates a vector with the classification methods that have
been tested to use with \link[methylearning]{apply_cls_method}.
}
\details{
While, in principle, all \link[caret]{train} "methods" would be
  suitable to use, only tested methods will be allowed, to ensure
  compatibility with \link[methylearning]{ml_cls} methods.
}
