% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{available_fs_methods}
\alias{available_fs_methods}
\title{Prints available feature selection methods}
\usage{
available_fs_methods(verbose = FALSE)
}
\arguments{
\item{verbose}{[logical] if TRUE, a summary list of the available methods is
returned. Default: FALSE.}
}
\value{
A list of three \code{character} vectors is returned: \code{all},
  with all the available feature selection methods; \code{filters}, with all
  the filter methods; \code{wrappers}, with all the wrapper methods.#'
}
\description{
This function returns a list with all the available feature selection
methods. If \code{verbose = TRUE}, a summary list of the available methods is
also printed.
}
\examples{
available_fs_methods()

}
