% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{boruta_fs}
\alias{boruta_fs}
\title{Boruta Algorithm for Feature Selection}
\usage{
boruta_fs(input_data, sample_labels, n, doTrace = 0)
}
\arguments{
\item{input_data}{[data.frame] a \code{data.frame} rows represent samples and
columns represent features (with beta values or methylation ratios).}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels.}

\item{n}{[numeric] the max number of features to output. It is also the
number of features in each GA "chromosome".}

\item{doTrace}{[numeric] controls \code{Boruta} verbosity. 0 means no
tracing, 1 means reporting decision about each attribute as soon as it is
justified, 2 means same as 1, plus reporting each importance source run.
Default = 0.}
}
\value{
[character] A character vector with \code{n} selected features.
}
\description{
This function implements a Boruta algorithm for feature selection. See
\code{Boruta::Boruta} documentation for further details.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
features_selected <- boruta_fs(input_data, sample_labels, 2)
print(features_selected)

}
\references{
Miron B. Kursa, Witold R. Rudnicki (2010). Feature Selection with
 the Boruta Package. Journal of Statistical Software, 36(11), p. 1-13.
 URL: \url{http://www.jstatsoft.org/v36/i11/}
}
