% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{cls_fitness}
\alias{cls_fitness}
\title{Calculate feature subset fitness}
\usage{
cls_fitness(subset, input_data, sample_labels)
}
\arguments{
\item{subset}{[character] a character vector of the feature names of the
particular subset to test accuracy.}

\item{input_data}{[data.frame] a \code{data.frame} rows represent samples and
columns represent features (with beta values or methylation ratios).}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels.}
}
\value{
the mean accuracy achieved by the three algorithms.
}
\description{
This function calculate feature subset fitness, based on classification
accuracy. Classification is performed using the three implemented algorithms:
 "knn", "C5.0" and "lda", averaging their results. Each of the algorithms is
performed with its default parameters.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
subset <- sample(colnames(input_data), 2)
accuracy_subset <- cls_fitness(subset, input_data, sample_labels)
print(accuracy_subset)

}
