% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{create_feature_list}
\alias{create_feature_list}
\title{Create the list of requested features from an ml_fs object}
\usage{
create_feature_list(ml_fs_obj, requested)
}
\arguments{
\item{ml_fs_obj}{[ml_fs] an object of the \code{ml_fs} class.}

\item{requested}{[character] a \code{character} vector with the feature
selection methods to be selected. Possible values: c("none", "all",
"filters", "wrappers", methylearning::selection_methods(ml_fs_obj)).
Values "all", "filters" and "wrappers" can only be indicated alone, or in
combination with "None". Value "none" means to use all the features,
regardless any feature selection method applied.}
}
\value{
A named \code{list} whith each feature selection.
}
\description{
This function creates a list of features requested, from the performed
feature selection of an ml_fs object. This function is meant to be used by
ml_cls constructor.
}
