% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\name{detect_platform}
\alias{detect_platform}
\title{Detect Illumina Array platform}
\usage{
detect_platform(gse)
}
\arguments{
\item{gse}{[GSE] a GSE object created with getGEO, containing an Illumina
array dataset.}
}
\value{
[character] A character vector containing either "27k", "450k" or
  "EPIC", depending on the correct platform detection or NULL if platform is
  not one of those.
}
\description{
This function returns either "27k", "450k", "EPIC" or NULL, depending on
the correct platform detection.
}
