% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cls_methods.R
\name{evaluation}
\alias{evaluation}
\title{Wrapper for predict + confusion matrix}
\usage{
evaluation(train_obj, test_data, test_labels)
}
\arguments{
\item{train_obj}{[train] an object of the \code{train} class, returned by
\link[caret]{train} function, or any other model object to be used with
\link[stats]{predict} function.}

\item{test_data}{[data.frame] data to be used as test data. It must contain
the same rows and columns as the data used to train the model.}

\item{test_labels}{[factor] sample observed (real) labels.}
}
\description{
This function is a wrapper that performs \link[stats]{predict} to predict
test labels from a test \code{data.frame} using a classification model. Then,
a \code{confusionMatrix} object is created using
\link[caret]{confusionMatrix}.
}
\examples{
set.seed(1234)
tr_dat <- data.frame(a = runif(10),
                     b = runif(10),
                     c = runif(10),
                     d = runif(10))
tr_lab <- factor(rbinom(10, 1, 0.5), labels = c("control", "treatment"))
# Train the model with this random data, using only 2 features.
model_cv <- apply_cls_method("knn", c("a", "b"), tr_dat, tr_lab)
# Evaluating in the same training data.
cm <- evaluation(model_cv, tr_dat[c("a", "b")], tr_lab)
print(cm)

}
