% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{ga_fs}
\alias{ga_fs}
\title{Wrapped Genetic Algorithm - knn/C5.0/lda for feature selection}
\usage{
ga_fs(input_data, sample_labels, n, cores, generations = 500, m_rate = 0.01,
  verbose = FALSE, return_history = FALSE)
}
\arguments{
\item{input_data}{[data.frame] a \code{data.frame} rows represent samples and
columns represent features (with beta values or methylation ratios).}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels.}

\item{n}{[numeric] the max number of features to output. It is also the
number of features in each GA "chromosome".}

\item{cores}{[numeric] number of cores to use in the parallel parts of the
algorithm.}

\item{generations}{[numeric] number of generations to perform. Default: 500.}

\item{m_rate}{[numeric] a number between 0 and 0.2. It will be taken as the
mutation rate. Default: 0.01.}

\item{verbose}{[logical] a logical indicating whether to status reports.
Default: FALSE.}

\item{return_history}{[logical] if \code{TRUE}, a list of two elements
(\code{$selection} and \code{$max_ac}) is returned. If \code{FALSE}, only
a vector with the selected features is returned. Default: FALSE.}
}
\value{
[character] A character vector with \code{n} selected features. If
  \code{return_history = TRUE}, a list is returned instead, with 3 elements:
  \code{$selection} [character] (the selected n features); \code{$max_ac}
  [numerical] a numerical vector with the max accuracy achieaved in each
  generation; \code{$mean_ac} [numerical] a numerical vector with the mean
  accuracy of th best 10% subsets selected in each generation.
}
\description{
This function implements a custom genetic algorithm wrapped to several
classification methods, to perform feature selection.
}
\details{
This is a very computational expensive feature selection method that
uses a genetic algorithm (GA) to select features. As fitness function to
optimize, this GA uses the mean accuracy obtained by the 3 implemented
classification algorithms (knn, C5.0 and lda). A recombination and mutation
step is performed in each generation.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
features_selected <- ga_fs(input_data, sample_labels, 50, 2)
print(features_selected)

}
