% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{ga_mutation}
\alias{ga_mutation}
\title{Generate random mutations for Genetic Algorithm}
\usage{
ga_mutation(subsets, remaining, m_rate = 0.01)
}
\arguments{
\item{subsets}{[list] a \code{list} of \code{character} vectors with the
subsets selected.}

\item{remaining}{[character] a \code{character} vector with the features not
selected.}

\item{m_rate}{[numeric] a number between 0 and 0.2. It will be taken as the
mutation rate. Default: 0.01.}
}
\value{
a \code{list} of two elements: the mutated \code{subsets list} and
  the mutated \code{remaining character} vector.
}
\description{
This function is intented to be called from ga_fs. Produce mutations, i.e.
feature swapping between remaining features and subsets, of each generation.
}
