% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{get_best_feature_set}
\alias{get_best_feature_set}
\title{Get Best Performant Feature Set Based on Training Validation}
\usage{
get_best_feature_set(ml_cls_obj)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.}
}
\value{
a \code{character} vector with the best performant feature set.
}
\description{
This function returns the best performant feature set based on training data
validation and metric used in \link[methylearning]{ml_cls} function
}
