% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{get_best_model_test}
\alias{get_best_model_test}
\title{Get Best Model From Test Evaluation}
\usage{
get_best_model_test(ml_cls_obj)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.}
}
\value{
a \code{list} with the following elements:
\describe{
  \item{best_model [train]}{a \code{train} object returned by
    \link[caret]{train}.}
  \item{best_fs [character]}{best feature selection method.}
  \item{best_cls [character]}{best classification method.}
  \item{best_acc [numeric]}{best accuracy obtained.}
  \item{best_kpp [numeric]}{best Kappa obtained.}
}
}
\description{
This function selects the best model based on test data evaluation, if
available. It returns a list with some objects (see Value).
}
