% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\name{get_chrXY}
\alias{get_chrXY}
\title{Get Sex Chromosome Probes}
\usage{
get_chrXY(platform)
}
\arguments{
\item{platform}{[character] a character vector of lenght 1, with either
"27k", "450k" or "EPIC".}
}
\value{
[character] A character vector all those probes located at sex
  chromosomes.
}
\description{
From a valid platform, this function returns a character vector with all the
probes located at sex chromosomes, based on platform annotation.
}
