% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{get_confusionMatrix_test.ml_cls}
\alias{get_confusionMatrix_test.ml_cls}
\title{Get Confusion Matrix From Test Evaluation}
\usage{
\method{get_confusionMatrix_test}{ml_cls}(ml_cls_obj, fs_method, cls_method,
  ...)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.}

\item{fs_method}{[character] an applied feature selection method.}

\item{cls_method}{[character] an applied classification method.}

\item{...}{additional arguments. None available yet...}
}
\description{
This function returns a \code{confusionMatrix} object from the selected
combination of \code{fs_method} and \code{cls_method}.
}
