% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{get_confusionMatrix_test}
\alias{get_confusionMatrix_test}
\title{Get Confusion Matrix From Test Evaluation}
\usage{
get_confusionMatrix_test(ml_cls_obj, ...)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.}

\item{...}{additional arguments for \link[methylearning]{get_confusionMatrix_test.ml_cls}
function.}
}
\value{
a \code{confusionMatrix} object, generated by
  \link[caret]{confusionMatrix}
}
\description{
This function returns a \code{confusionMatrix} object from the selected
combination of \code{fs_method} and \code{cls_method}.
}
