% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\name{get_GEO_methylarray}
\alias{get_GEO_methylarray}
\title{Prepare Methylation Array Data From GEO}
\usage{
get_GEO_methylarray(gse, target_name, exclude_sex_probes = TRUE,
  exclude_cross_reactive = TRUE, exclude_polymorphic = TRUE,
  limma_filter = FALSE, random_filter = FALSE)
}
\arguments{
\item{gse}{[GSE] an object of \code{GSE} class, created using
\code{GEOquery::getGEO}.}

\item{target_name}{[character] a character vector of the target (class
labels) column name in \code{gse phenoData data.frame}.}

\item{exclude_sex_probes}{[logical] should chrX and chrY probes be removed?
Default: TRUE.}

\item{exclude_cross_reactive}{[logical] should cross-reactive probes be
removed? Default: TRUE.}

\item{exclude_polymorphic}{[logical] should polymorphic (i.e. probes with
known SNP at CpG position) be removed? Default: TRUE.}

\item{limma_filter}{[logical/numeric] should limma filtering be performed?
Default: FALSE. If an integer is specified, this will be the number of
features selected.}

\item{random_filter}{[logical/numeric] should random filtering be performed?
Default: FALSE. If an integer is specified, this will be the number of
features selected}
}
\value{
[ml_data] an object of ml_data class.
}
\description{
This function uses a \code{GSE} object (a \code{list} object) created with
\link[GEOquery]{getGEO}, containing only methylation array data,  and process
it to create an \code{ml_data} object.
}
\details{
By default data is preprocessed to remove sex chromosome,
  cross-reactive (multi-mapping) and polymorphic (with known SNP) probes.
  Moreover, one of th following filters can be applied to reduce the number
  of features (probes): limma_filter, based on limma statistical
  significance; random_filter, selecting random probes.
}
\examples{
\dontrun{
if (require(GEOquery) &
    require(Biobase)) {
  gse <- getGEO("GSE69229")
  ml <- get_GEO_methylarray(gse = gse, target_name = "outcome:ch1")
  summary(ml)
}
}

}
