% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{get_test_results}
\alias{get_test_results}
\title{Get Test Results for Accuracy and Kappa}
\usage{
get_test_results(ml_cls_obj)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.}
}
\value{
a \code{list} with two elements named "Accuracy" and "Kappa", each
  containing a \code{data.frame} with the obtained values, where rows are
  feature selection methods and columns are classification methods.
}
\description{
Get a list with two \code{data.frame}, one for accuracy and the other for
kappa, where rows are feature selection methods and columns are
classification methods.
}
