% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{get_training_results}
\alias{get_training_results}
\title{Get Training Results for Accuracy and Kappa}
\usage{
get_training_results(ml_cls_obj)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.}
}
\description{
Get a list with two \code{data.frame}, one for accuracy and the other for
kappa, where rows are feature selection methods and columns are
classification methods.
}
