% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{information_gain_fs}
\alias{information_gain_fs}
\title{Entropy-based information gain feature selection}
\usage{
information_gain_fs(input_data, sample_labels, n)
}
\arguments{
\item{input_data}{[data.frame] a \code{data.frame} rows represent samples and
columns represent features (with beta values or methylation ratios).}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels.}

\item{n}{[numeric] the max number of features to output.}
}
\value{
[character] A character vector with at most \code{n} selected
  features.
}
\description{
This funtion uses FSelector::information.gain implementation to select
features based on entropy information gain between feature and
\code{sample_labels}.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
features_selected <- information_gain_fs(input_data, sample_labels, 2)
print(features_selected)

}
