% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{limma_fs}
\alias{limma_fs}
\title{Select CpGs by differential methylation using limma}
\usage{
limma_fs(input_data, sample_labels, n)
}
\arguments{
\item{input_data}{[data.frame] a \code{data.frame} rows represent samples and
columns represent features (beta values or methylation ratios).}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels.}

\item{n}{[numeric] the number of features (top differentially methylated CpGs
to select).}
}
\value{
[character] A character vector with the \code{n} top differentially
  methylated (sorted by p-value) feature names, for all contrasts tested.
  For more than two groups in \code{sample_labels}, each contrast is designed
  as one Vs the others.
}
\description{
This funtion uses limma to select top differentially methylated probes
(features) between groups (sample labels factor). The number of class labels
can be 2 or more, but no paired designs are implemented. It is equivalent to
anova_fs, but much faster. Only applicable to methylation array data.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
features_selected <- limma_fs(input_data, sample_labels, 2)
print(features_selected)

}
