% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data.R
\name{ml_data}
\alias{ml_data}
\title{Creates \code{ml_data} objects}
\usage{
ml_data(input, labels_column = 1)
}
\arguments{
\item{input}{[data.frame] input \code{data.frame} with samples as rows and
features as columns. There must be an extra column with class labels.}

\item{labels_column}{[numeric] a number indicating the column index of the
class label variable.}
}
\value{
An object of [ml_data] class.
}
\description{
Constructor for \code{ml_data} objects. This object encapsulates input data
to be used with rest of package functions.
}
\details{
Object members:
\describe{
\item{classes [character]}{class labels present in the data set.}
\item{feature_types [character]}{object classes of feature variables.}
\item{sample_names [character]}{all samples names. They are always
  \code{input data.frame} \code{rownames}.}
\item{training_data [data.frame]}{training data set (only features, not
  labels). Samples as rows and features as columns.}
\item{training_labels [factor]}{training class labels.}
\item{training_sample_names [character]}{training sample names.}
\item{test_data [data.frame]}{test data set (only features). It is only
  present when \code{data.partitioned} is TRUE, otherwise NULL.}
\item{test_labels [factor]}{test class labels. It is only present when
  \code{data.partitioned} is TRUE, otherwise NULL.}
\item{test_sample_labels [character]}{test sample names. It is only
  present when \code{data.partitioned} is TRUE, otherwise NULL.}
\item{data_partitioned [logic]}{TRUE when \code{test_samples} are
  provided.}
}

Object methods:
[summary]
[split_data]
}
\examples{
set.seed(1234)
df <- data.frame(a = runif(10, 0, 1),
                 b = runif(10, 0, 1),
                 c = as.factor(as.integer(runif(10, 0, 2))))
data <- ml_data(df, 3)
summary(data)

}
