% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{mRMR_fs}
\alias{mRMR_fs}
\title{Minimum redundancy, maximum relevance feature filtering}
\usage{
mRMR_fs(input_data, sample_labels, n, cores)
}
\arguments{
\item{input_data}{[data.frame] a \code{data.frame} rows represent samples and
columns represent features (with beta values or methylation ratios).}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels.}

\item{n}{[numeric] the max number of features to output.}

\item{cores}{[numeric] number of cores to use.}
}
\value{
[character] A character vector with at most \code{n} selected
  features.
}
\description{
This function implements mRMRe::mRMR.classic algorithm for feature selection.
The algorithm finds weights of continous and discrete attributes basing on a
distance between instances.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
features_selected <- relief_fs(input_data, sample_labels, 2)
print(features_selected)

}
