% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{plot.ml_cls}
\alias{plot.ml_cls}
\title{Plot function for 'ml_cls' objects}
\usage{
\method{plot}{ml_cls}(x, mode = "Accuracy", fs_ROC = "", cls_ROC = "", time_mode = "sum", ...)
}
\arguments{
\item{x}{[ml_cls] an object of 'ml_cls' class.}

\item{mode}{[character] one of the following strings: c("Accuracy", "Kappa",
"ROC", "Time") to indicate the type of graphic to draw. Default: "Accuracy".}

\item{fs_ROC}{[character] a feature selection method previously performed, to
construct ROC curve. When mode = "ROC", this should be a valid method.}

\item{cls_ROC}{[character] a classification method previously performed, to
construct ROC curve. When mode = "ROC", this should be a valid method.}

\item{time_mode}{[character] whether to plot classification times only
(time_mode = 'cls_only') or to plot the sum of feature selection and
classification times (time_mode = 'sum'). Default: 'sum'.}

\item{...}{additional arguments to pass to \link[graphics]{plot}.}
}
\value{
a plot.
}
\description{
This function plot some graphics from a 'ml_cls' object, based on training
validation.
}
\details{
This function can draw four type of graphics all of them based
  on training validation:
  \describe{
    \item{Accuracy barplot}{Barplot with accuracy values and accuracy sd.}
    \item{Kappa barplot}{Barplot with Kappa values and Kappa sd.}
    \item{ROC curves}{ROC curves and AUC values. For two-classes
      classification problems.}
    \item{Computation time barplot}{Barplot with the computation time took
      by classification or by feature selection + classification.}
  }
}
