% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fs.R
\name{plot_fs_time}
\alias{plot_fs_time}
\title{Plot Computation Time for \code{ml_fs} object}
\usage{
plot_fs_time(fs_object)
}
\arguments{
\item{fs_object}{[ml_fs] an object of \code{ml_fs} class.}
}
\value{
a barchart with the time elapsed in each feature selection
  computation.
}
\description{
This function creates a plot of the computation time elapsed in each of the
feature selection calculations performed on a \code{ml_fs} object.
}
\examples{
set.seed(1234)
df <- data.frame(a = runif(10, 0, 1),
                 b = runif(10, 0, 1),
                 c = as.factor(as.integer(runif(10, 0, 2))))
data <- ml_data(df, 3)
data_sel <- ml_fs(data, fs_methods = "filters", selection_size = 2)
plot_fs_time(data_sel)

}
