% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{plot_test.ml_cls}
\alias{plot_test.ml_cls}
\title{Specialized plot function for test data evaluation}
\usage{
\method{plot_test}{ml_cls}(ml_cls_obj, mode = "Accuracy", ...)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.}

\item{mode}{[character] one of the following strings: c("Accuracy", "Kappa")
to indicate the type of graphic to draw.}

\item{...}{additional arguments. None available yet.}
}
\description{
This function plot 'Accuracy' or 'Kappa' barcharts from an 'ml_cls' object,
based on test data evaluation.
}
\details{
This function can draw two types of graphics all of them based
  on test data evaluation:
  \describe{
    \item{Accuracy barplot}{Barplot with accuracy values.}
    \item{Kappa barplot}{Barplot with Kappa values.}
  }
}
