% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{plot_test}
\alias{plot_test}
\title{Specialized plot function for test data evaluation}
\usage{
plot_test(ml_cls_obj, ...)
}
\arguments{
\item{ml_cls_obj}{[ml_cls] an object of 'ml_cls' class.#'}

\item{...}{additional arguments for \link[methylearning]{plot_test.ml_cls}.}
}
\value{
a plot.
}
\description{
This function plot 'Accuracy' or 'Kappa' barcharts from an 'ml_cls' object,
based on test data evaluation.
}
\details{
This function can draw two types of graphics all of them based
  on test data evaluation:
  \describe{
    \item{Accuracy barplot}{Barplot with accuracy values.}
    \item{Kappa barplot}{Barplot with Kappa values.}
  }
}
