% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fs.R
\name{plot_venn_fs.ml_fs}
\alias{plot_venn_fs.ml_fs}
\title{Plot Venn Diagram of Feature Selection Results}
\usage{
\method{plot_venn_fs}{ml_fs}(fs_object, fs_methods, ...)
}
\arguments{
\item{fs_object}{[ml_fs] an object of \code{ml_fs} class.}

\item{fs_methods}{[character] a \code{character} vector with the name of
the feature selection methods to compare. From 1 to 5 methods are allowed.}

\item{...}{additional optional arguments for
\link[VennDiagram]{draw.pairwise.venn},
\link[VennDiagram]{draw.triple.venn},
\link[VennDiagram]{draw.quad.venn} or
\link[VennDiagram]{draw.quintuple.venn}.}
}
\description{
This function uses \link[VennDiagram]{VennDiagram} package to plot Venn
diagrams of the Feature Selection results, for up to 5 methods.
}
\details{
For 2 methods comparison \link[VennDiagram]{draw.pairwise.venn}
  is used. For 3 methods comparison \link[VennDiagram]{draw.triple.venn}
  is used. For 4 methods comparison \link[VennDiagram]{draw.quad.venn}
  is used. For 5 methods comparison \link[VennDiagram]{draw.quintuple.venn}
  is used. Additional arguments for each of the
  \link[VennDiagram]{VennDiagram} functions are available.
}
