% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fs.R
\name{plot_venn_fs}
\alias{plot_venn_fs}
\title{Plot Venn Diagram of Feature Selection Results}
\usage{
plot_venn_fs(fs_object, ...)
}
\arguments{
\item{fs_object}{[ml_fs] an object of \code{ml_fs} class.}

\item{...}{additional arguments for \link[methylearning]{plot_venn_fs.ml_fs}.}
}
\value{
a Venn diagram plot.
}
\description{
This function uses \link[VennDiagram]{VennDiagram} package to plot Venn
diagrams of the Feature Selection results, for up to 5 methods.
}
\details{
For 2 methods comparison \link[VennDiagram]{draw.pairwise.venn}
  is used. For 3 methods comparison \link[VennDiagram]{draw.triple.venn}
  is used. For 4 methods comparison \link[VennDiagram]{draw.quad.venn}
  is used. For 5 methods comparison \link[VennDiagram]{draw.quintuple.venn}
  is used. Additional arguments for each of the
  \link[VennDiagram]{VennDiagram} functions are available.
}
\examples{
set.seed(1234)
df <- data.frame(a = runif(10, 0, 1),
                 b = runif(10, 0, 1),
                 c = as.factor(as.integer(runif(10, 0, 2))))
data <- ml_data(df, 3)
data_sel <- ml_fs(data, fs_methods = "filters", selection_size = 2)
plot_venn_fs(data_sel, fs_methods = c("anova_fs", "mRMR_fs"),
             fill = c("red", "blue"), category = c("anova_fs", "mRMR_fs"))

}
