% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\docType{data}
\name{polymorphic450k}
\alias{polymorphic450k}
\title{Documentation for datasets used by get_GEO_methylarray}
\description{
Documentation for datasets used by get_GEO_methylarray
}
\keyword{data}
