% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\docType{data}
\name{probesxy27k}
\alias{probesxy27k}
\title{Documentation for datasets used by get_GEO_methylarray}
\description{
27k chrX and chrY probes based on:
IlluminaHumanMethylation27k.db::IlluminaHumanMethylation27kCHR
}
\keyword{data}
