% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\docType{data}
\name{probesxy450k}
\alias{probesxy450k}
\title{Documentation for datasets used by get_GEO_methylarray}
\description{
450k chrX and chrY probes based on:
IlluminaHumanMethylation450kanno.ilmn12.hg19
}
\keyword{data}
