% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GEO_methylarray.R
\docType{data}
\name{probesxyEPIC}
\alias{probesxyEPIC}
\title{Documentation for datasets used by get_GEO_methylarray}
\description{
EPIC chrX and chrY probes based on:
IlluminaHumanMethylationEPICanno.ilm10b4.hg19
}
\keyword{data}
