% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{random_fs}
\alias{random_fs}
\title{Random Feature Selection}
\usage{
random_fs(features, n)
}
\arguments{
\item{features}{[character] a \code{character} vector with the data
\code{colnames}, i.e. feature names (probes or CpGs).}

\item{n}{[numeric] the number of features to select.}
}
\value{
[character] A character vector with the features \code{colnames}
  selected.
}
\description{
Implementation of random feature selection, for testing and performance
evaluation.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
features <- colnames(input_data)
features_selected <- random_fs(features, 2)
print(features_selected)

}
