% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_methods.R
\name{relief_fs}
\alias{relief_fs}
\title{RELIEF algorithm for feature selection}
\usage{
relief_fs(input_data, sample_labels, n, neighbours_count = 5,
  sample_size = 10)
}
\arguments{
\item{input_data}{[data.frame] a \code{data.frame} rows represent samples and
columns represent features (with beta values or methylation ratios).}

\item{sample_labels}{[factor] a \code{factor} of length == nrow(input_data),
containing the sample labels.}

\item{n}{[numeric] the max number of features to output.}

\item{neighbours_count}{[numeric] number of neighbours to find for every
sampled instance. Default: 5.}

\item{sample_size}{[numeric] number of instances to sample. Default: 10.}
}
\value{
[character] A character vector with at most \code{n} selected
  features.
}
\description{
This function implements FSelector::relief algorithm for feature selection.
The algorithm finds weights of continous and discrete attributes basing on a
distance between instances.
}
\examples{
set.seed(1234)
input_data <- data.frame(a = runif(10, min = 0, max = 1),
                         b = runif(10, min = 0, max = 1),
                         c = runif(10, min = 0, max = 1),
                         d = runif(10, min = 0, max = 1))
sample_labels <- as.factor(c(rep("classA", 5), rep("classB", 5)))
features_selected <- relief_fs(input_data, sample_labels, 2)
print(features_selected)

}
