% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fs.R
\name{selection_methods}
\alias{selection_methods}
\title{Get Feature Selection methods used}
\usage{
selection_methods(fs_object)
}
\arguments{
\item{fs_object}{[ml_fs] an object of \code{ml_fs}
class.}
}
\value{
A \code{character} vector with the used feature selection methods.
}
\description{
This function returns the feature selection methods already used to create
an object of the \code{ml_fs} class.
}
\examples{
set.seed(1234)
df <- data.frame(a = runif(10, 0, 1),
                 b = runif(10, 0, 1),
                 c = as.factor(as.integer(runif(10, 0, 2))))
data <- ml_data(df, 3)
data_sel <- ml_fs(data, fs_methods = "filters", selection_size = 2)
m <- selection_methods(data_sel)
print(m)

}
