% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fs.R
\name{selection_results}
\alias{selection_results}
\title{Get Feature Selection Results}
\usage{
selection_results(fs_object)
}
\arguments{
\item{fs_object}{[ml_fs] an object of \code{ml_fs}
class.}
}
\value{
a \code{list} with the \code{character} vector of the selected
  features, for each of the used methods.
}
\description{
This function returns the results of previous feature selection, stored in an
object of the \code{ml_fs} class.
}
\examples{
set.seed(1234)
df <- data.frame(a = runif(10, 0, 1),
                 b = runif(10, 0, 1),
                 c = as.factor(as.integer(runif(10, 0, 2))))
data <- ml_data(df, 3)
data_sel <- ml_fs(data, fs_methods = "filters", selection_size = 2)
m <- selection_results(data_sel)
print(m)

}
