% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data.R
\name{split_data}
\alias{split_data}
\title{Splits data into training and data set.}
\usage{
split_data(ml_obj, splitting)
}
\arguments{
\item{ml_obj}{[ml_data] an object of the class \code{ml_data}.}

\item{splitting}{[numeric] an optional numeric to express the proportion
of cases to be used in training data. Default: 2/3.}
}
\value{
the original ml_obj, but with data partitioned.
}
\description{
Method only available for objects of the \code{ml_data} class. It splits
the actual training data of an unpartitioned \code{ml_data} object into
training data and test data, using caret::createDataPartition().
}
\examples{
\dontrun{
set.seed(1234)
if (require(GEOquery) &
    require(Biobase)) {
  gse <- getGEO("GSE69229")
  ml <- get_GEO_methylarray(gse = gse, target_name = "outcome:ch1")
  summary(ml)
  ml_partitioned <- split_data(ml)
  summary(ml_partitioned)
}
}

}
