% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_cls.R
\name{summary.ml_cls}
\alias{summary.ml_cls}
\title{Prints a summary for a \code{ml_cls} object}
\usage{
\method{summary}{ml_cls}(object, ...)
}
\arguments{
\item{object}{[ml_cls_obj] an object of the \code{ml_cls} class.}

\item{...}{additional arguments to pass to \code{summary.default}. See
\link[base]{summary}.}
}
\value{
a printed summary.
}
\description{
\code{summary} method for objects of the \code{ml_cls} class.
}
\examples{
set.seed(1234)
\dontrun{
if (require(GEOquery) &
    require(Biobase)) {
  gse <- getGEO("GSE69229")
  # Generate ml_data object.
  ml <- get_GEO_methylarray(gse = gse, target_name = "outcome:ch1")
  # Generate ml_fs object.
  ml_f <- ml_fs(ml, fs_methods = "anova_fs", selection_size = 50)
  # Generate ml_cls object.
  ml_c <- ml_cls(ml_f, cls_methods = "knn")
  summary(ml_c)
}
}

}
