% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_data.R
\name{summary.ml_data}
\alias{summary.ml_data}
\title{Prints a summary for a \code{ml_data} object}
\usage{
\method{summary}{ml_data}(object, ...)
}
\arguments{
\item{object}{[ml_data] an object of the class \code{ml_data}.}

\item{...}{additional arguments to pass to \code{summary.default}. See
\link[base]{summary}.}
}
\value{
a printed summary.
}
\description{
\code{summary} method for objects of the class \code{ml_data}.
}
\examples{
set.seed(1234)
df <- data.frame(a = runif(10, 0, 1),
                 b = runif(10, 0, 1),
                 c = as.factor(as.integer(runif(10, 0, 2))))
data <- ml_data(df, 3)
summary(data)

}
