% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_fs.R
\name{summary.ml_fs}
\alias{summary.ml_fs}
\title{Prints a summary for a \code{ml_fs} object}
\usage{
\method{summary}{ml_fs}(object, ...)
}
\arguments{
\item{object}{[ml_fs] an object of the class
\code{ml_fs}.}

\item{...}{additional arguments to pass to \code{summary.default}. See
\link[base]{summary}.}
}
\value{
a printed summary.
}
\description{
\code{summary} method for objects of the class \code{ml_fs}. It
prints all the information inherited from \code{ml_data} class, plus the
specific \code{ml_fs} data.
}
\examples{
set.seed(1234)
df <- data.frame(a = runif(10, 0, 1),
                 b = runif(10, 0, 1),
                 c = as.factor(as.integer(runif(10, 0, 2))))
data <- ml_data(df, 3)
data_sel <- ml_fs(data, fs_methods = "anova_fs", selection_size = 2)
summary(data_sel)

}
